refactor: compartmentalize router files
