+++ web/src/main.js
@@ -4,7 +4,7 @@ import { createApp } from 'vue'
 import { createPinia } from 'pinia'
 
 import App from './App.vue'
-import router from './router'
+import router from './router/_index'
 
 import 'vuetify/styles'
 import { createVuetify } from 'vuetify'
