--- web/src/router/index.js
@@ -1,48 +1,0 @@
-import { createRouter, createWebHistory } from 'vue-router'
-import LoginView from '../views/LoginView.vue'
-import CreateAccountView from '../views/CreateAccountView.vue'
-import DashboardView from '../views/users/DashboardView.vue'
-import TeamsView from '../views/teams/TeamsView.vue'
-import TeamView from '../views/teams/TeamView.vue'
-
-const router = createRouter({
-  history: createWebHistory(import.meta.env.BASE_URL),
-  routes: [
-    {
-      path: '/login',
-      name: 'login',
-      component: LoginView
-    },
-    {
-      path: '/about',
-      name: 'about',
-      // route level code-splitting
-      // this generates a separate chunk (About.[hash].js) for this route
-      // which is lazy-loaded when the route is visited.
-      component: () => import('../views/AboutView.vue')
-    },
-    {
-      path: '/create-account',
-      name: 'create-account',
-      component: CreateAccountView
-    },
-    {
-      path: '/dashboard',
-      name: 'dashboard',
-      component: DashboardView
-    },
-    {
-      path: '/teams',
-      name: 'teams',
-      component: TeamsView
-    },
-    {
-      path: '/teams/:teamId',
-      name: 'team',
-      component: TeamView,
-      props: true
-    }
-  ]
-})
-
-export default router
