+++ web/src/router/shifts.js
@@ -1,0 +1,5 @@
+const shifts = [
+
+]
+
+export { shifts }
