+++ web/src/router/users.js
@@ -1,0 +1,5 @@
+const users = [
+
+]
+
+export { users }
